# webui/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strategy Research API",
    "description": "HTTP API for quantnodes strategy research framework",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/admin/compaction/keep-all/{enabled}": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Toggle Keep All Compactions",
        "description": "Toggle keep_all_compactions_in_history filter at runtime.\n\nArgs:\n    enabled: true = force all compactions in LLM history (legacy);\n             false = filter to most recent only (new behavior).\n    confirm: must be 'yes' to prevent accidental toggles.\n\nReturns:\n    Status of the toggle.",
        "operationId": "toggle_keep_all_compactions_api_admin_compaction_keep_all__enabled__post",
        "parameters": [
          {
            "name": "enabled",
            "in": "path",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Enabled"
            }
          },
          {
            "name": "confirm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Must be 'yes' to confirm",
              "default": "",
              "title": "Confirm"
            },
            "description": "Must be 'yes' to confirm"
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toggle Keep All Compactions Api Admin Compaction Keep All  Enabled  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/audit-log": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Audit Log",
        "description": "Get recent admin actions (max 100, FIFO).",
        "operationId": "get_audit_log_api_admin_audit_log_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Audit Log Api Admin Audit Log Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/hangs/report": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Hangs Report",
        "description": "Aggregate hanging-protection events for the ops runbook.\n\nEvents are recorded by each protection layer (LLM wall-clock\ntimeout, backtest log stall, agent no_progress, circuit breaker\nopen, watchdog interrupt) into the goals DB ``hanging_events``\ntable. This endpoint reports how often each fired in the window —\nthe C.2 daily-drive signal for threshold tuning.",
        "operationId": "get_hangs_report_api_admin_hangs_report_get",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 720,
              "minimum": 0.1,
              "description": "Look-back window in hours",
              "default": 24,
              "title": "Hours"
            },
            "description": "Look-back window in hours"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Hangs Report Api Admin Hangs Report Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}